Reorganise utils
